refactor: split UserService - extract ProfileImageService and align to hex ports
- Extract updateProfileImage() and GCS logic into ProfileImageService (application/service/)
- Add LoadUserPort methods: findByEmail, findByRefreshToken
- Add SaveUserPort output port interface
- Update UserPersistenceAdapter to implement LoadUserPort + SaveUserPort
- Refactor UserService to use ports instead of direct UserRepository access (240 lines)
- Update UserController to inject ProfileImageService for profile image endpoint
- Update UserServiceTest to mock ports instead of UserRepository
- Fix pre-existing test compilation errors in DocumentApiIntegrationTest and CompanyApiIntegrationTest
